--- patches/st-no_bold_colors-20160727-308bfbf.diff
@@ -1,13 +1,0 @@
-diff --git a/st.c b/st.c
-index 2594c65..f6fe717 100644
---- a/st.c
-+++ b/st.c
-@@ -3719,7 +3719,7 @@ xdrawglyphfontspecs(const XftGlyphFontSpec *specs, Glyph base, int len, int x, i
- 
- 	/* Change basic system colors [0-7] to bright system colors [8-15] */
- 	if ((base.mode & ATTR_BOLD_FAINT) == ATTR_BOLD && BETWEEN(base.fg, 0, 7))
--		fg = &dc.col[base.fg + 8];
-+		fg = &dc.col[base.fg];
- 
- 	if (IS_SET(MODE_REVERSE)) {
- 		if (fg == &dc.col[defaultfg]) {
