+++ patches/st-scrollback-mouse-0.8.diff
@@ -1,8 +1,8 @@
 diff --git a/config.def.h b/config.def.h
-index eae969e..34ebb44 100644
+index 27d42ca..feec7e2 100644
 --- a/config.def.h
 +++ b/config.def.h
-@@ -152,8 +152,14 @@ static unsigned int defaultunderline = 7;
+@@ -156,8 +156,14 @@ static unsigned int defaultattr = 11;
   */
  static MouseShortcut mshortcuts[] = {
  	/* button               mask            string */
@@ -12,18 +12,18 @@ index eae969e..34ebb44 100644
 +	{ Button5,              XK_NO_MOD,      "\005" },
 +};
 +
-+static MouseKey mkeys[] = {
++MouseKey mkeys[] = {
 +	/* button               mask            function        argument */
 +	{ Button4,              ShiftMask,      kscrollup,      {.i =  1} },
 +	{ Button5,              ShiftMask,      kscrolldown,    {.i =  1} },
  };
  
  /* Internal keyboard shortcuts. */
-diff --git a/st.c b/st.c
-index 233d301..c71b6e7 100644
---- a/st.c
-+++ b/st.c
-@@ -299,6 +299,13 @@ typedef union {
+diff --git a/st.h b/st.h
+index 628e876..cdd25ae 100644
+--- a/st.h
++++ b/st.h
+@@ -81,6 +81,13 @@ typedef union {
  	const void *v;
  } Arg;
  
@@ -34,18 +34,27 @@ index 233d301..c71b6e7 100644
 +	const Arg arg;
 +} MouseKey;
 +
- typedef struct {
- 	uint mod;
- 	KeySym keysym;
-@@ -953,6 +960,7 @@ bpress(XEvent *e)
+ void die(const char *, ...);
+ void redraw(void);
+ void draw(void);
+@@ -129,3 +136,4 @@ extern char *termname;
+ extern unsigned int tabspaces;
+ extern unsigned int defaultfg;
+ extern unsigned int defaultbg;
++extern MouseKey mkeys[];
+diff --git a/x.c b/x.c
+index d43a529..754d859 100644
+--- a/x.c
++++ b/x.c
+@@ -409,6 +409,7 @@ bpress(XEvent *e)
  {
  	struct timespec now;
  	MouseShortcut *ms;
 +	MouseKey *mk;
+ 	int snap;
  
  	if (IS_SET(MODE_MOUSE) && !(e->xbutton.state & forceselmod)) {
- 		mousereport(e);
-@@ -967,6 +975,14 @@ bpress(XEvent *e)
+@@ -424,6 +425,14 @@ bpress(XEvent *e)
  		}
  	}
  
@@ -58,5 +67,5 @@ index 233d301..c71b6e7 100644
 +	}
 +
  	if (e->xbutton.button == Button1) {
- 		clock_gettime(CLOCK_MONOTONIC, &now);
- 
+ 		/*
+ 		 * If the user clicks below predefined timeouts specific
