+++ run-patches.sh
@@ -1,0 +1,8 @@
+#!/bin/sh
+
+patch < patches/st-clipboard-20180309-c5ba9c0.diff 
+patch < patches/st-no_bold_colors-0.8.1.diff 
+patch < patches/st-scrollback-0.8.diff 
+patch < patches/st-scrollback-mouse-0.8.diff 
+patch < patches/st-vertcenter-20180320-6ac8c8a.diff
+patch < patches/st-alpha-0.8.1.diff
