+++ st-remove.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+
+rm /usr/local/bin/st
+rm /usr/local/share/man/man1/st.1
+rm /usr/share/terminfo/s/st*
