+++ st.1
@@ -120,7 +120,7 @@ st executes
 instead of the shell.  If this is used it
 .B must be the last option
 on the command line, as in xterm / rxvt.
-This option is only intended for compability,
+This option is only intended for compatibility,
 and all the remaining arguments are used as a command
 even without it.
 .SH SHORTCUTS
@@ -142,28 +142,25 @@ Print the full screen to the
 Print the selection to the
 .I iofile.
 .TP
-.B Alt-Shift-Page Up
+.B Ctrl-Shift-Page Up
 Increase font size.
 .TP
-.B Alt-Shift-Page Down
+.B Ctrl-Shift-Page Down
 Decrease font size.
 .TP
-.B Alt-Shift-Home
+.B Ctrl-Shift-Home
 Reset to default font size.
 .TP
-.B Shift-Insert
+.B Ctrl-Shift-y
 Paste from primary selection (middle mouse button).
 .TP
-.B Alt-Shift-Insert
-Paste from clipboard selection.
-.TP
-.B Alt-Shift-c
+.B Ctrl-Shift-c
 Copy the selected text to the clipboard selection.
 .TP
-.B Alt-Shift-v
+.B Ctrl-Shift-v
 Paste from the clipboard selection.
 .TP
-.B Alt-Ctrl
+.B Ctrl-Shift-i
 Launch dmenu to enter a unicode codepoint and send the corresponding glyph
 to st.
 .SH CUSTOMIZATION
