+++ st.info
@@ -25,6 +25,7 @@ st| simpleterm,
 	cuu=\E[%p1%dA,
 	dch=\E[%p1%dP,
 	dch1=\E[P,
+	dim=\E[2m,
 	dl=\E[%p1%dM,
 	dl1=\E[M,
 	ech=\E[%p1%dX,
@@ -142,7 +143,6 @@ st| simpleterm,
 	khome=\E[1~,
 	kil1=\E[2;5~,
 	krmir=\E[2;2~,
-	kich1=\E[2~,
 	knp=\E[6~,
 	kmous=\E[M,
 	kpp=\E[5~,
@@ -185,16 +185,23 @@ st| simpleterm,
 	tsl=\E]0;,
 	xenl,
 	vpa=\E[%i%p1%dd,
-# Tmux unofficial extensions, see TERMINFO EXTENSIONS in tmux(1)
+# XTerm extensions
+	rmxx=\E[29m,
+	smxx=\E[9m,
+# tmux extensions, see TERMINFO EXTENSIONS in tmux(1)
 	Se,
 	Ss,
 	Tc,
+	Ms=\E]52;%p1%s;%p2%s\007,
 
 st-256color| simpleterm with 256 colors,
 	use=st,
+	ccc,
 	colors#256,
+	oc=\E]104\007,
 	pairs#32767,
 #	Nicked from xterm-256color
+	initc=\E]4;%p1%d;rgb\:%p2%{255}%*%{1000}%/%2.2X/%p3%{255}%*%{1000}%/%2.2X/%p4%{255}%*%{1000}%/%2.2X\E\\,
 	setab=\E[%?%p1%{8}%<%t4%p1%d%e%p1%{16}%<%t10%p1%{8}%-%d%e48;5;%p1%d%;m,
 	setaf=\E[%?%p1%{8}%<%t3%p1%d%e%p1%{16}%<%t9%p1%{8}%-%d%e38;5;%p1%d%;m,
